% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DividendCalc_sim.R
\name{DividendCalc_sim}
\alias{DividendCalc_sim}
\title{Simulate Smoke-free Dividend Calculation}
\usage{
DividendCalc_sim(
  data_spend,
  data_prevalence,
  upshift = 1.417,
  div = 0.93,
  illicit_prop = 1298/(14307 + 1298),
  n_sim = 100,
  seed = 2021
)
}
\arguments{
\item{data_spend}{Data table. Weekly spending by local authority from the cleaned toolkit data.}

\item{data_prevalence}{Data table. Smoking prevalence by local authority from OHID tobacco profiles.}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}

\item{div}{Numeric. Proportion of licit tobacco expenditure which is smoke-free dividend.}

\item{illicit_prop}{Numeric. The proportion of total tobacco expenditure which is illicit.}

\item{n_sim}{Numeric. Number of simulations.}

\item{seed}{Numeric. Random number seed.}
}
\value{
Smoke-free dividend calculations with simulated uncertainty
}
\description{
Calculate the smoke-free dividend for each local authority, using simulation
methods to incorporate parameter uncertainty in weekly expenditure and smoking prevalence
}
\examples{

\dontrun{

}
}
