% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFingertips.R
\name{GetFingertips}
\alias{GetFingertips}
\title{Get Smoking Prevalence Data}
\usage{
GetFingertips(year = 2022)
}
\arguments{
\item{year}{Numeric. year of smoking prevalence data to obtain.}
}
\value{
A data frame of smoking prevalence by UTLA for a single year
}
\description{
A function which uses the fingertipsR package to obtain and clean data on
smoking prevalence by local authority from the OHID local tobacco profiles. By default
data are returned for the upper tier local authority boundaries as at December 2019.
}
\examples{

\dontrun{

}

}
