% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ltla_2012_to_utla_2019}
\alias{ltla_2012_to_utla_2019}
\title{Mapping 2012 LTLA to 2019 UTLA}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 326 rows and 4 columns.
}
\usage{
ltla_2012_to_utla_2019
}
\description{
A lookup table of correspondence between lower tier local authority (LTLA) identifiers used
in the Smoking Toolkit Study data and the upper tier local authority (UTLA) identifiers for
local authorities as at December 2019
}
\keyword{datasets}
