% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_utla_2019_2020}
\alias{mapping_utla_2019_2020}
\title{Mapping UTLAs (2019 - 2020)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 151 rows and 6 columns.
}
\usage{
mapping_utla_2019_2020
}
\description{
A lookup table for mapping the April 2019 - March 2020 UTLAs to the April 2020 - March 2021 UTLAs
}
\keyword{datasets}
