% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_utla_2020_2022}
\alias{mapping_utla_2020_2022}
\title{Mapping UTLAs (2020 - 2022)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 152 rows and 6 columns.
}
\usage{
mapping_utla_2020_2022
}
\description{
A lookup table for mapping the April 2020 - March 2021 UTLAs to the April 2021 - March 2023 UTLAs
}
\keyword{datasets}
