% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_counts}
\alias{population_counts}
\title{Population counts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 152 rows and 10 columns.
}
\usage{
population_counts
}
\description{
Mid-year population estimates by local authority (2022 UTLAs) for years 2015-2021. Data obtained
from the [Office for National Statistics](https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland).
}
\details{
Data accessed on 15/12/2023
}
\keyword{datasets}
