% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_to_utla}
\alias{region_to_utla}
\title{Region to local authority lookup}
\format{
An object of class \code{data.frame} with 152 rows and 3 columns.
}
\usage{
region_to_utla
}
\description{
Mapping of 2022 UTLAs to their regions
}
\keyword{datasets}
