% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smkfreediv2.R
\name{smkfreediv2}
\alias{smkfreediv2}
\title{Smoke-free Dividend}
\usage{
smkfreediv2(
  year = 2022,
  read_toolkit = TRUE,
  in_path,
  out_path,
  data,
  name,
  start_month = 90,
  end_month = 160,
  index = smkfreediv2::cpi_tobacco,
  base_month = 12,
  base_year = 2018,
  upshift = 1.417,
  div = 0.93,
  illicit_prop = 1298/(14307 + 1298),
  n_sim = 100,
  seed = 42
)
}
\arguments{
\item{year}{Numeric. year of smoking prevalence data to use in the calculations.}

\item{read_toolkit}{Logical. If TRUE, read in raw toolkit SPSS data, if FALSE use toolkit data already created using `ToolkitRead()`.
If option is set to FALSE, the toolkit data will be read from the directory set by `out_path` and use the file
name specified by `name`.}

\item{in_path}{Character. path to the directory where the raw SPSS data is saved.}

\item{out_path}{Character. Path to the directory where the working version of the data in RDS format will be exported.}

\item{data}{name of the raw toolkit data file (defaults to the April 2021 version).}

\item{name}{Character. Name of the file to export.}

\item{start_month}{Numeric. Toolkit tart month for the analysis period (default = 90, April 2014).}

\item{end_month}{Numeric. Toolkit end month for the analysis period (default = 160, February 2020).}

\item{index}{Data table. Tobacco price data to use in the deflation. Default is the ONS tobacco CPI series.}

\item{base_month}{Numeric. Month of the year to deflate currency variables to. (Default = 12).}

\item{base_year}{Numeric. Year to deflate currency variables to. (Default = 2018).}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}

\item{div}{Numeric. Proportion of licit tobacco expenditure which is smoke-free dividend.}

\item{illicit_prop}{Numeric. The proportion of total tobacco expenditure which is illicit.}

\item{n_sim}{Numeric. Number of simulations.}

\item{seed}{Numeric. Random number seed.}
}
\value{
Smoke-free dividend calculations with simulated uncertainty
}
\description{
Wrapper function for obtaining data, calculating the smoke-free dividend for England local areas, defined
as the 152 upper-tier local authorities in 2022, and simulating the uncertainty around the estimates.
}
\examples{

\dontrun{

}
}
