% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolkitCalcWeekSpend.R
\name{ToolkitCalcWeekSpend}
\alias{ToolkitCalcWeekSpend}
\title{Weekly Tobacco Spend}
\usage{
ToolkitCalcWeekSpend(data, upshift = 1.417)
}
\arguments{
\item{data}{Data table. The cleaned toolkit data.}

\item{upshift}{Numeric. Parameter to upshift expenditures in the toolkit data to account for under-reporting.}
}
\value{
mean and median weekly spending and standard errors by subgroup
}
\description{
Generates weekly spending on tobacco as a weighted mean across individuals in the sample.
}
\examples{

\dontrun{

}
}
