% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolkitClean.R
\name{ToolkitClean}
\alias{ToolkitClean}
\title{Clean Toolkit Data}
\usage{
ToolkitClean(
  data,
  start_month = 90,
  end_month = 160,
  index = smkfreediv2::cpi_tobacco,
  base_month = 12,
  base_year = 2018
)
}
\arguments{
\item{data}{Data table. Toolkit data.}

\item{start_month}{Numeric. Toolkit tart month for the analysis period (default = 90, April 2014).}

\item{end_month}{Numeric. Toolkit end month for the analysis period (default = 160, February 2020).}

\item{index}{Data table. Tobacco price data to use in the deflation. Default is the ONS tobacco CPI series.}

\item{base_month}{Numeric. Month of the year to deflate currency variables to. (Default = 12).}

\item{base_year}{Numeric. Year to deflate currency variables to. (Default = 2018).}
}
\value{
Cleaned toolkit data
}
\description{
Clean the raw toolkit data to produce a dataset that can be used for the
smoke free dividend calculations
}
\examples{

\dontrun{

}
}
