% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolkitRead.R
\name{ToolkitRead}
\alias{ToolkitRead}
\title{Read Toolkit Data}
\usage{
ToolkitRead(in_path, out_path, data, save = TRUE, name)
}
\arguments{
\item{in_path}{path to the directory where the raw SPSS data is saved.}

\item{out_path}{path to the directory where the working version of the data will be exported.}

\item{data}{name of the data file (defaults to the April 2021 version)}

\item{save}{Logical. If TRUE, exports out a copy of the data in RDS format to the directory specified by `out_path`}

\item{name}{Character. Name of the file to export}
}
\value{
Data frame of STS data
}
\description{
Reads in the raw Smoking Toolkit Study (STS) data in SPSS format
and saves to a chosen directory in RDS format as well as returning it as
an R object
}
\examples{

\dontrun{

}
}
