% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utla_2019_20}
\alias{utla_2019_20}
\title{Local Authority Lookup 2019-20}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 151 rows and 2 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 151 rows and 2 columns.
}
\usage{
utla_2019_20

utla_2019_20
}
\description{
Names and area codes of the 151 upper tier local authorities (UTLAs) in England April 2019 - March 2020

Names and area codes of the 151 upper tier local authorities (UTLAs) in England April 2019 - March 2020
}
\keyword{datasets}
