% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utla_2021_23}
\alias{utla_2021_23}
\title{Local Authority Lookup 2020-21}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 152 rows and 2 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 152 rows and 2 columns.
}
\usage{
utla_2021_23

utla_2021_23
}
\description{
Names and area codes of the 151 upper tier local authorities (UTLAs) in England April 2020 - March 2021

Names and area codes of the 152 upper tier local authorities (UTLAs) in England April 2021 - March 2023
}
\keyword{datasets}
