% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utla_2023}
\alias{utla_2023}
\title{Local Authority Lookup 2023}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 153 rows and 2 columns.
}
\usage{
utla_2023
}
\description{
Names and area codes of the 153 upper tier local authorities (UTLAs) in England from April 2023
}
\keyword{datasets}
